// FarmSolution/FarmZone.Frontend/src/app/client/apiclient.ts
export interface AnimalDto {
    externalId?: string;
    name?: string | undefined;
}